Update index.js
Let me know how retarded the thing I just did is. Trying to wrap the cards[i].card_title in a regexp to mark them as they're listed in the json as case insensitive and then mirror the == check of message.text to be case insensitive. LOLOLOL I'm expecting it to totally be wrong somehow.